fix(core): short-circuit identical compare refs
